+++ src/main/java/se/bjurr/prnfb/listener/PrnfbRenderer.java
@@ -13,9 +13,9 @@ import static se.bjurr.prnfb.listener.PrnfbRenderer.REPO_PROTOCOL.http;
 import static se.bjurr.prnfb.listener.PrnfbRenderer.REPO_PROTOCOL.ssh;
 
 import java.io.UnsupportedEncodingException;
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.List;
 import java.util.logging.Logger;
 import java.util.regex.Matcher;
 
@@ -448,14 +448,15 @@ public class PrnfbRenderer {
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
      Map<PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore, boolean shouldAcceptAnyCertificate) {
     List<PullRequestParticipant> slist = newArrayList(pullRequest.getReviewers());
-    StringBuilder rString = new StringBuilder();
 
-    String sep = ",";
-    for (PullRequestParticipant each : slist) {
-        rString.append(sep).append(each.getUser().getDisplayName());
+    if (slist.isEmpty()) {
+     return "";
     }
-
-    return rString.substring(1);
+    StringBuilder sb = new StringBuilder();
+    for (PullRequestParticipant each : slist) {
+     sb.append("," + each.getUser().getDisplayName());
+    }
+    return sb.substring(1);
    }
   }), PULL_REQUEST_REVIEWERS_ID(new Resolver() {
    @Override
@@ -464,14 +465,15 @@ public class PrnfbRenderer {
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
      Map<PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore, boolean shouldAcceptAnyCertificate) {
     List<PullRequestParticipant> slist = newArrayList(pullRequest.getReviewers());
-    StringBuilder rString = new StringBuilder();
 
-    String sep = ",";
-    for (PullRequestParticipant each : slist) {
-        rString.append(sep).append(Integer.toString(each.getUser().getId()));
+    if (slist.isEmpty()) {
+     return "";
     }
-
-    return rString.substring(1);
+    StringBuilder sb = new StringBuilder();
+    for (PullRequestParticipant each : slist) {
+     sb.append("," + each.getUser().getId());
+    }
+    return sb.substring(1);
    }
   }), PULL_REQUEST_REVIEWERS_SLUG(new Resolver() {
    @Override
@@ -480,14 +482,15 @@ public class PrnfbRenderer {
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
      Map<PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore, boolean shouldAcceptAnyCertificate) {
     List<PullRequestParticipant> slist = newArrayList(pullRequest.getReviewers());
-    StringBuilder rString = new StringBuilder();
 
-    String sep = ",";
-    for (PullRequestParticipant each : slist) {
-        rString.append(sep).append(each.getUser().getSlug());
+    if (slist.isEmpty()) {
+     return "";
     }
-
-    return rString.substring(1);
+    StringBuilder sb = new StringBuilder();
+    for (PullRequestParticipant each : slist) {
+     sb.append("," + each.getUser().getSlug());
+    }
+    return sb.substring(1);
    }
   }), PULL_REQUEST_REVIEWERS_APPROVED_COUNT(new Resolver() {
    @Override